+++ Classes/Extensions.Identity.Scopes.cs
@@ -7,9 +7,7 @@
 /// https://github.com/cjvandyk/Extensions/blob/main/LICENSE
 /// </summary>
 
-using System.Security.Cryptography.X509Certificates;
-
-namespace System.Identity
+namespace Extensions.Identity
 {
     /// <summary>
     /// A class containing the various scope string arrays.